Handler 객체 분리
- Handler 인터페이스를 구현하고
- LIst에 담아서 사용할 수 있도록 함
