+++ src/main/java/webserver/Mapping/MappingMatcher.java
@@ -1,22 +1,29 @@
 package webserver.Mapping;
 
-import webserver.HttpHandler.RequestHandler;
+import webserver.HttpHandler.Handler;
+import webserver.HttpHandler.LoginHandler;
+import webserver.HttpHandler.ResourceHandler;
+import webserver.HttpHandler.UserHandler;
 import webserver.HttpMessage.Request;
 import webserver.HttpMessage.RequestStartLine;
 import webserver.HttpMessage.Response;
 
 import java.lang.reflect.Method;
+import java.util.List;
 import java.util.function.BiPredicate;
 
 public class MappingMatcher {
     private final Request request;
     private final RequestStartLine startLine;
-    private final RequestHandler handler;
+    private final List<Handler> handlers = List.of(
+            new LoginHandler(),
+            new ResourceHandler(),
+            new UserHandler()
+    );
 
     public MappingMatcher(Request request) {
         this.request = request;
         this.startLine = request.getStartLine();
-        handler = new RequestHandler();
     }
 
     public Response getResponse() throws Exception {
@@ -24,21 +31,23 @@ public class MappingMatcher {
         final String path = startLine.getUri().split("\\?")[0];
 
         if (httpMethod.equals("GET")) {
-            return handleRequest(path, handler, this::matchGetMapping);
+            return handleRequest(path, this::matchGetMapping);
         }
         if (httpMethod.equals("POST")) {
-            return handleRequest(path, handler, this::matchPostMapping);
+            return handleRequest(path, this::matchPostMapping);
         } else throw new IllegalAccessException("설정되어 있지 않은 http 메소드입니다.");
     }
 
-    private Response handleRequest(String uri, RequestHandler handler, BiPredicate<Method, String> matchMapping) throws Exception {
-        for (Method method : RequestHandler.class.getDeclaredMethods()) {
-            if (matchMapping.test(method, uri)) {
-                return (Response) method.invoke(handler, request);
+    private Response handleRequest(String uri, BiPredicate<Method, String> matchMapping) throws Exception {
+        for(Handler handler : handlers) {
+            for (Method method : handler.getClass().getDeclaredMethods()) {
+                if (matchMapping.test(method, uri)) {
+                    return (Response) method.invoke(handler, request);
+                }
             }
         }
         // default : get Resource
-        return handleRequest("/" , handler , matchMapping);
+        return handleRequest("/" , matchMapping);
     }
 
     private boolean matchGetMapping(Method method, String uri) {
